Update troubleshoot-hybrid-join-windows-current.md
In Step 4 (Discover Phase), users are instructed to test if URLs are accessible from the device in system context.  As we do have a script for this referenced at https://learn.microsoft.com/en-us/samples/azure-samples/testdeviceregconnectivity/testdeviceregconnectivity/. it is a good place to add this script, so users do not have to spend time looking for a way to access mentioned URLs in System context.